Reverted the dropping of conditional/noop attrs
If it's convenient, _and_ saves a bit of code, I don't really see the
reasoning to drop it.

At least added an LFSR_ATTR_IF  macro to try to reduce the C expression
noise.

Also I'm keeping the explicit noop grows in lfsr_btree_commit. I think
it's a good decision to prevent accidental non-nooping in the future. If
these grows didn't noop, it wouldn't be an error, but a minor parasitic
drain on performance/wear. Not great...

            code          stack
  before:  34052           2928
  after:   33940 (-0.3%)   2928 (+0.0%)