Limited block_size and in-block types to 28-bits
One downside of leb128 encoding is that the worst case encoded size is
not that well aligned due to a relatively underutilized last byte:

  0xffffffff => 0xff 0xff 0xff 0xff 0x0f

This normally doesn't really matter, the whole point of leb128 is that
larger encodings are statistically less likely. But in littlefs we need
to allocate the worst-case buffer size in order to encode/decode
leb128s, and these buffers need to stick around on the stack during
metadata commit calls, which are also the point of highest stack usage
in the system.

But 32-bits is somewhat arbitrary, it just happens to be our register
size. In fact, we're not really using 32-bits, but instead only 31-bits
to take advantage of the sign bit for ad-hoc sum types:

  0x7fffffff => 0xff 0xff 0xff 0xff 0x07

In theory, if we limit this further to 28-bits, we could save some stack
space:

  0x7fffffff => 0xff 0xff 0xff 0xff 0x07
  0x0fffffff => 0xff 0xff 0xff 0x7f

This may seem like a small amount of savings, but it also restores
alignment to the encoding, and should result in less wasted padding
around buffers.

Though it's important to note these are the most valuable bits, as the
range grows exponentially with each bit added. Reducing 31-bits to
28-bits reduces the range from ~2GiB to ~256MiB:

  0x7fffffff => 2,147,483,647
  0x0fffffff =>   268,435,455

---

At the moment I'm hesistant to reduce _all_ on-disk leb128s to 28-bits.

The signed-32-bit limit of ~2GiB is fairly well understood in this
space, mainly thanks to FAT, and reducing this to ~256MiB risks quite a
surprise to users (it's also a regression from the current littlefs
version).

But one type where this limit is pretty reasonable is our block_size.

I don't think we'll see devices with erase blocks >256MiB for a while,
and at the very least those devices will probably need a 64-bit
filesystem for other reasons anyways...

And limiting block_size to <=256MiB has a surprising number of knock-on
effects:

- The tag size/jump field never exceeds 28-bits, reducing worst-case tag
  dsize from 12 bytes -> 11 bytes.

  The also reduces our worst-case attr-estimate from 40 bytes ->
  37 bytes

- rbyd/btree trunks never exceed 28-bits, saving space in shrub/branch/
  btree encodings.

- The bptr encoding is reduced from 24 bytes -> 21 bytes, since several
  of its fields are in-block (size, off, cksize).

- The commit checksum encoding is reduced by a byte for every commit,
  from 12 bytes -> 11 bytes.

  This is due to needing to expand the cksum tag's size field to the
  worst possible leb128 encoding due to a catch-22 situation.

Unfortunately the actual stack savings is a bit underwhelming:

            code          stack
  before:  33688           2808
  after:   33700 (+0.0%)   2800 (-0.3%)

This may be because, by adopting 28-bits in only some fields, most
buffers still end up unaligned and the on-stack size doesn't change due
to padding. Or it could just be that I'm overestimating the cost of our
on-stack buffers.

Still, I think the change is worth keeping if only for the reducing
attr-estimate and saved byte on every on-disk commit.

In the future it would be interesting to explore additional
configurations, e.g. a 28-bit flavor of littlefs to compliment this
31-bit flavor. You could imagine the fitting into other register sized
flavors for different capacity/code cost/device compat tradeoffs:

  flavor               register  leb128   size-limit
  14-bit littlefs  =>  16-bit    2 bytes  ~16KiB
  15-bit littlefs  =>  16-bit    3 bytes  ~32KiB
  28-bit littlefs  =>  32-bit    4 bytes  ~256MiB
  31-bit littlefs  =>  32-bit    5 bytes  ~2GiB
  56-bit littlefs  =>  64-bit    8 bytes  ~64PiB
  63-bit littlefs  =>  64-bit    9 bytes  ~8ExiB

This is where the on-disk size-limit attr would really shine.

---

Note we don't need an additional on-disk limit attr for the block_size.
We already store the block_size in the superblock, so we just need to
error if attempting to mount a filesystem with block_size >256MiB.