Reduce stack allocation in the little-leb128 decoder
This avoids the extra stack allocation for the unaligned worst-case
leb128 encoding by duplicating most of the "big-leb128" decoder. The
upside is less stack usage, but at a code cost, since we basically have
two copies of this function now.

This is a bit of a tough call, the percentage change is basically the
same:
            code          stack
  before:  33700           2800
  after:   33808 (+0.3%)   2792 (-0.3%)

On one hand, we would want to deduplicate these functions if they end up
with the same encoding cost (28-bit littlefs mode?), and less code is
less code, on the other hand, RAM is in general more valuable than
code...

This may be worth reverting in the future...